Move `.types` into `.models`
