+++ src/flickr_api/__init__.py
@@ -11,29 +11,6 @@ from .exceptions import (
     UnrecognisedFlickrApiException,
     UserDeleted,
 )
-from .types import (
-    Comment,
-    CommonsInstitution,
-    MachineTags,
-    License,
-    LicenseId,
-    User,
-    UserInfo,
-    TakenGranularity,
-    DateTaken,
-    LocationInfo,
-    Size,
-    SafetyLevel,
-    SinglePhoto,
-    AlbumContext,
-    GalleryContext,
-    GroupContext,
-    MediaType,
-    PhotoContext,
-    SinglePhotoInfo,
-    Tag,
-    Visibility,
-)
 
 
 __version__ = "2.21.0"
@@ -47,30 +24,10 @@ __all__ = [
     "InvalidApiKey",
     "InvalidXmlException",
     "LicenseNotFound",
-    "License",
-    "LicenseId",
-    "LocationInfo",
-    "MachineTags",
-    "User",
-    "UserInfo",
-    "TakenGranularity",
-    "Comment",
-    "CommonsInstitution",
-    "DateTaken",
-    "Size",
-    "SafetyLevel",
-    "SinglePhoto",
-    "SinglePhotoInfo",
-    "AlbumContext",
-    "GalleryContext",
-    "GroupContext",
     "InsufficientPermissionsToComment",
-    "MediaType",
     "PhotoContext",
     "PhotoIsPrivate",
-    "Tag",
     "UnrecognisedFlickrApiException",
     "UserDeleted",
-    "Visibility",
     "__version__",
 ]
