+++ src/flickr_api/models/licenses.py
