+++ src/flickr_api/models/machine_tags.py
