+++ src/flickr_api/models/sizes.py
