+++ src/flickr_api/utils.py
@@ -4,7 +4,7 @@ Convert values retrieved from the Flickr API into nicely-typed values.
 
 from xml.etree import ElementTree as ET
 
-from .types import LocationInfo, SafetyLevel
+from .models import LocationInfo, SafetyLevel
 
 
 def parse_safety_level(s: str) -> SafetyLevel:
